terminer all configuration jwt and refresh token

+++ CMspringboot/src/main/java/ma/youcode/cmspringboot/repository/UserRepository.java
@@ -7,8 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-
-public interface UserRepository extends JpaRepository<AppUser, Long> {
+public interface UserRepository extends JpaRepository<AppUser, Integer> {
     Optional<AppUser> findByUsername(String username);
     Optional<AppUser> findByNum(Integer user_num);
 }
